refactor: split out caching
