+++ packages/myst-execute/src/index.ts
@@ -1,2 +1,4 @@
 export { kernelExecutionTransform } from './execute.js';
 export { findExistingJupyterServer, launchJupyterServer } from './manager.js';
+export { LocalDiskCache } from './cache.js';
+export type { ICache } from './cache.js';
